[spinel] move `SpiFrame` implementation to spi_frame.hpp
