+++ src/lib/spinel/BUILD.gn
@@ -36,6 +36,7 @@ spinel_sources = [
   "openthread-spinel-config.h",
   "radio_spinel.hpp",
   "radio_spinel_impl.hpp",
+  "spi_frame.hpp",
   "spinel.c",
   "spinel_buffer.cpp",
   "spinel_buffer.hpp",
