+++ src/posix/platform/spi_interface.hpp
@@ -37,12 +37,12 @@
 #include "openthread-posix-config.h"
 
 #include "platform-posix.h"
+#include "lib/hdlc/hdlc.hpp"
 #include "lib/spinel/multi_frame_buffer.hpp"
+#include "lib/spinel/spi_frame.hpp"
 #include "lib/spinel/spinel_interface.hpp"
 
 #include <openthread/openthread-system.h>
-
-#include "ncp/ncp_spi.hpp"
 
 namespace ot {
 namespace Posix {
